added new property
