+++ submission/submission-config/src/main/kotlin/ac/uk/ebi/biostd/common/properties/ClusterProperties.kt
@@ -11,6 +11,21 @@ data class ClusterProperties(
     val enabled: Boolean,
     val user: String,
     val key: String,
-    val server: String,
+    val lsfServer: String,
     val logsPath: String,
+    val default: Cluster,
 )
+
+enum class Cluster {
+    LSF,
+    ;
+
+    companion object {
+        fun fromName(name: String): Cluster {
+            return when (name.uppercase()) {
+                "LSF" -> LSF
+                else -> throw IllegalArgumentException("$name is not a valid cluster name")
+            }
+        }
+    }
+}
