+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/cluster/web/ClusterExecutor.kt
@@ -1,6 +1,7 @@
 package ac.uk.ebi.biostd.cluster.web
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
+import ac.uk.ebi.biostd.common.properties.Cluster
 import uk.ac.ebi.biostd.client.cluster.api.LsfClusterClient
 import uk.ac.ebi.biostd.client.cluster.model.Job
 import uk.ac.ebi.biostd.client.cluster.model.JobSpec
@@ -9,7 +10,7 @@ class ClusterExecutor private constructor(val lsfClusterClient: LsfClusterClient
     constructor(properties: ApplicationProperties) : this(
         LsfClusterClient.create(
             properties.cluster.key,
-            properties.cluster.server,
+            properties.cluster.lsfServer,
             properties.cluster.logsPath,
         ),
     )
@@ -47,20 +48,6 @@ class ClusterExecutor private constructor(val lsfClusterClient: LsfClusterClient
     ): String {
         return when (cluster) {
             Cluster.LSF -> lsfClusterClient.jobLogs(jobId)
-        }
-    }
-}
-
-enum class Cluster {
-    LSF,
-    ;
-
-    companion object {
-        fun fromName(name: String): Cluster {
-            return when (name.uppercase()) {
-                "LSF" -> LSF
-                else -> throw IllegalArgumentException("$name is not a valid cluster name")
-            }
         }
     }
 }
